move backfill logic to on fetch
