+++ packages/server/src/sdk/plugins/mapping.json
